# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: courier-outbox
  version: 0.3.0
  description: |
    Reliable outbound webhook delivery service.

    This document lists the endpoints that exist today. Each feature change
    must update it in the same PR as the handler (see
    docs/architecture/patterns-layering.md). The planned surface is in
    SRS §4.2 and openspec/specs/.

    Every path under `/v1` requires `Authorization: Bearer <api_key>`.
    Subscription CRUD and event enqueue are implemented. Delivery and DLQ
    are not.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
security: []
paths:
  /healthz:
    get:
      operationId: getHealthz
      summary: Liveness probe
      tags: [ops]
      security: []
      responses:
        "200":
          description: Process is alive.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [ok]
  /readyz:
    get:
      operationId: getReadyz
      summary: Readiness probe
      description: |
        Pings Postgres and Redis. Returns 200 when both answer, otherwise
        503 with `code` = `unavailable` naming the failing dependencies.
        Driver errors are never included in the response.
      tags: [ops]
      security: []
      responses:
        "200":
          description: Dependencies are reachable.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [ready]
        "503":
          description: One or more dependencies did not answer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: unavailable
                message: "not ready: redis"
  /v1/subscriptions:
    get:
      operationId: listSubscriptions
      summary: List subscriptions
      description: Returns subscriptions owned by the authenticated API key. Signing secrets are never included.
      tags: [subscriptions]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The caller's subscriptions, oldest first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createSubscription
      summary: Create a subscription
      description: |
        Registers a webhook endpoint. The signing secret is returned once
        as standard base64 of 32 random bytes and is stored encrypted.
      tags: [subscriptions]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionRequest"
      responses:
        "201":
          description: Subscription created. Copy `signing_secret` now.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionWithSecret"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
  /v1/subscriptions/{id}:
    parameters:
      - $ref: "#/components/parameters/SubscriptionId"
    get:
      operationId: getSubscription
      summary: Get a subscription
      tags: [subscriptions]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The subscription. `signing_secret` is omitted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateSubscription
      summary: Update a subscription
      description: |
        Partial update. `rotate_secret=true` issues a new signing secret
        and returns it once. A foreign or missing id is 404.
      tags: [subscriptions]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchSubscriptionRequest"
      responses:
        "200":
          description: Updated subscription. `signing_secret` is present only after rotate.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionWithSecret"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
  /v1/subscriptions/{id}/disable:
    parameters:
      - $ref: "#/components/parameters/SubscriptionId"
    post:
      operationId: disableSubscription
      summary: Disable a subscription
      description: Sets `enabled` to false. Idempotent. The row remains readable.
      tags: [subscriptions]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The disabled subscription.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/events:
    post:
      operationId: enqueueEvent
      summary: Enqueue an event
      description: |
        Persists the event and one pending delivery per matching enabled
        subscription in a single transaction. Reusing `idempotency_key`
        for the same API key returns 200 with `Idempotent-Replay: true`
        and the original ids. Request bodies larger than 256 KiB are 413.
      tags: [events]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnqueueEventRequest"
      responses:
        "201":
          description: Event created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnqueueEventResponse"
        "200":
          description: Idempotent replay of an existing event.
          headers:
            Idempotent-Replay:
              required: true
              schema:
                type: string
                enum: ["true"]
              description: Always `true` on a replay.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnqueueEventResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
  /v1/events/{id}:
    parameters:
      - $ref: "#/components/parameters/EventId"
    get:
      operationId: getEvent
      summary: Get an event
      description: |
        Returns event metadata, payload, and delivery summaries for the
        calling API key. A foreign or missing id is 404.
      tags: [events]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The event owned by the caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: |
        Producer API key printed once by `courier keys create --name <name>`.
        Format `co_<prefix>_<secret>`. Send as `Authorization: Bearer <key>`.
  parameters:
    SubscriptionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Subscription id. Malformed values return 400, not 404.
    EventId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Event id. Malformed values return 400, not 404.
  responses:
    Unauthorized:
      description: |
        Missing, malformed, unknown, wrong, or revoked API key. All five
        cases share this status, body, and `WWW-Authenticate` header.
      headers:
        WWW-Authenticate:
          schema:
            type: string
            example: 'Bearer realm="courier"'
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: unauthorized
            message: missing or invalid api key
    NotFound:
      description: Resource does not exist or is owned by another API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: not_found
            message: resource not found
    ValidationFailed:
      description: Malformed JSON, unknown field, or domain validation failure.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: validation_failed
            message: "target_url: must use https"
    PayloadTooLarge:
      description: Request body exceeded 256 KiB.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: payload_too_large
            message: request body too large
  schemas:
    Error:
      type: object
      description: Error envelope returned by every non-2xx response.
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable snake_case identifier.
          example: not_found
        message:
          type: string
          description: Human-readable, safe to display. Never contains secrets.
          example: resource not found
    Subscription:
      type: object
      required: [id, target_url, event_types, headers, enabled, description, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        target_url:
          type: string
          format: uri
        event_types:
          type: array
          items:
            type: string
        headers:
          type: object
          additionalProperties:
            type: string
        enabled:
          type: boolean
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SubscriptionWithSecret:
      allOf:
        - $ref: "#/components/schemas/Subscription"
        - type: object
          properties:
            signing_secret:
              type: string
              description: Standard base64 of 32 random bytes. Present only on create and rotate.
    SubscriptionList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Subscription"
    CreateSubscriptionRequest:
      type: object
      required: [target_url]
      properties:
        target_url:
          type: string
          format: uri
        event_types:
          type: array
          items:
            type: string
          default: ["*"]
        headers:
          type: object
          additionalProperties:
            type: string
        description:
          type: string
    PatchSubscriptionRequest:
      type: object
      additionalProperties: false
      properties:
        target_url:
          type: string
          format: uri
        event_types:
          type: array
          items:
            type: string
        headers:
          type: object
          additionalProperties:
            type: string
        description:
          type: string
        enabled:
          type: boolean
        rotate_secret:
          type: boolean
    EnqueueEventRequest:
      type: object
      additionalProperties: false
      required: [type, payload, idempotency_key]
      properties:
        type:
          type: string
          minLength: 1
          maxLength: 128
          pattern: '^[a-z0-9]+([._-][a-z0-9]+)*$'
        payload:
          type: object
          additionalProperties: true
          description: JSON object. The whole request body must be at most 256 KiB.
        idempotency_key:
          type: string
          minLength: 1
          maxLength: 128
        occurred_at:
          type: string
          format: date-time
          description: RFC 3339. Defaults to the server time when omitted.
    EnqueueEventResponse:
      type: object
      required: [event_id, delivery_ids]
      properties:
        event_id:
          type: string
          format: uuid
        delivery_ids:
          type: array
          items:
            type: string
            format: uuid
    Event:
      type: object
      required: [id, type, payload, idempotency_key, occurred_at, created_at, deliveries]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        payload:
          type: object
          additionalProperties: true
        idempotency_key:
          type: string
        occurred_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        deliveries:
          type: array
          items:
            $ref: "#/components/schemas/DeliverySummary"
    DeliverySummary:
      type: object
      required: [id, subscription_id, status, attempt_count, next_attempt_at]
      properties:
        id:
          type: string
          format: uuid
        subscription_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, retrying, delivered, dead_lettered]
        attempt_count:
          type: integer
          minimum: 0
        next_attempt_at:
          type: string
          format: date-time